Swapped querySelectors for useRefs. Moved scroll function into parent components.
