+++ src/components/Header.jsx
@@ -3,27 +3,13 @@ import logo from  '../assets/italianlogo.png';
 import { useNavigate } from "react-router-dom";
 
 
-const Header = () => {
-  const navigate = useNavigate();
-
-  const handleNavLogoClick = (event) => {
-    event.preventDefault();
-    
-    const recipeSection = document.querySelector(".detail-section");
-    recipeSection.setAttribute('style', 'opacity: 0; transition: opacity 0.4s;');
-
-    setTimeout(() => {
-      window.scrollTo({
-        top: 0,
-        behavior: "smooth"
-      });
-      navigate(`/`);
-    }, 400);
-  }
+const Header = ({navigateToHome}) => {
 
   return (
     <header>
-      <Link onClick={handleNavLogoClick} className="nav-logo">
+      <Link 
+        className="nav-logo"
+        onClick={navigateToHome}>
         <img src={logo} alt='logo' className='logo' />
       </Link>
     </header>
