+++ src/view/Home.jsx
@@ -2,7 +2,7 @@ import useContentful from '../hook/useContentful';
 import { useLoaderData, useNavigate } from "react-router-dom";
 import RecipeCard from '../components/RecipeCard';
 import RecipeList from '../components/RecipeList';
-import React, { useState, useEffect } from 'react';
+import React, { useState, useEffect, useRef } from 'react';
 
 export function loader() {
   const { getReceipes } = useContentful();
@@ -11,31 +11,37 @@ export function loader() {
 
 function Home() {
   const receipes = useLoaderData()
-  const [loading, setLoading] = useState(true);
+  const navigate = useNavigate();
 
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setLoading(false);
-    }, 1000); 
-  }, []);
+  const ref = useRef(null);
+  const scrollToTop = (elementId) => {
+    ref.current.setAttribute('style', 'opacity: 0; transition: opacity 0.3s;')
 
-  if (loading) {
-    return (
-      <div className="loadingDiv">
-        <div className="loaderAnimation"></div>
-      </div>
-    );
+    window.scrollTo({
+        top: 0,
+        behavior: "smooth"
+    });
+
+    setTimeout(() => {
+        navigate(`/${elementId}`);
+    }, 300);
+
+    ref.current.setAttribute('style', 'opacity: 1')
+
   };
-  
+
+
   return (
     <>
-      <section>
+      <section ref={ref}>
 
         <RecipeCard
-        recipes={receipes} />
+        recipes={receipes}
+        scrollToTop={scrollToTop}/>
 
         <RecipeList
-        recipes={receipes} />
+        recipes={receipes}
+        scrollToTop={scrollToTop} />
 
       </section>  
     </>
